requirements.txt
aiogram==2.9
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
multidict==4.7.6
pytz==2020.1
typing-extensions==3.7.4.3
yarl==1.5.1
SQLAlchemy>=1.0.12
SQLAlchemy-Utils>=0.32.0
python-dotenv==0.14.0